+++ Components/src/components/Display/DisplayPokemon.tsx
@@ -1,6 +1,6 @@
 import { Component } from 'react';
-import NamedApiResource from '../types/namedAPIResource';
-import Pokemon from '../types/pokemon';
+import NamedApiResource from '../../types/namedAPIResource';
+import Pokemon from '../../types/pokemon';
 
 type Props = {
   pokemon: NamedApiResource;
@@ -43,7 +43,7 @@ class DisplayPokemon extends Component<Props> {
 
     return (
       pokemon && (
-        <div key={pokemon.id}>
+        <div key={pokemon.id} className="pokemon">
           <b>{pokemon.name}</b>
           <img
             src={pokemon.sprites.front_default}
