+++ Components/src/components/ErrorBoundary/ErrorPopUp.tsx
@@ -1,5 +1,6 @@
 import { Component, ErrorInfo } from 'react';
 import helpers from '../../helpers';
+import './ErrorPopUp.css';
 
 type Props = {
   error: Error | null;
@@ -12,14 +13,15 @@ class ErorrPopUp extends Component<Props> {
   };
 
   render() {
+    helpers.saveSearchText('');
+    helpers.saveSearchType('');
+
     return (
       <div className={`error${this.state.isVisible ? ' visible' : ''}`}>
         <h2>An error occured! Check console for more info!</h2>
         <button
           onClick={() => {
             this.setState({ isVisible: false });
-            helpers.saveSearchText('');
-            helpers.saveSearchType('');
           }}
         >
           Close popup
