+++ Components/src/components/Search/SearchTerms.tsx
